Remove internal export leak: relink submodules to real owners
Relocate the last cross-cutting helpers out of core.ts (dedupeNonEmpty,
reddit error classifiers, global-blocklist readers, getCurrentSubredditNameCompat)
and repoint every submodule import at its owning module. core.ts is now
a pure re-export barrel exporting exactly the original 115 public names
(0 internal helpers leaked). All 521 declarations remain byte-identical;
typecheck, 167 tests, and build pass.
